requirements.txt
python-dotenv
hugchat
